icode

+++ leetcode/169_Majority_Element.h
@@ -1,16 +1,25 @@
 #include "../base/icode.h"
-class L169:public icode {
+class L169 : public icode
+{
 public:
-    int majorityElement(vector<int>& nums) {
-        int cur=nums[0], time=1, i=1;
-        for (; i<nums.size(); ++i) {
-            if (time <= 0) {
+    int majorityElement(vector<int> &nums)
+    {
+        int cur = nums[0], time = 1, i = 1;
+        for (; i < (int)nums.size(); ++i)
+        {
+            if (time <= 0)
+            {
                 cur = nums[i];
                 time = 1;
-            } else {
-                if (nums[i] == cur) {
+            }
+            else
+            {
+                if (nums[i] == cur)
+                {
                     ++time;
-                } else {
+                }
+                else
+                {
                     --time;
                 }
             }
@@ -18,8 +27,9 @@ public:
         return cur;
     }
 
-    void run() {
-        vector<int> nums = {2,2,1,1,1,2,2};
+    void run()
+    {
+        vector<int> nums = {2, 2, 1, 1, 1, 2, 2};
         cout << majorityElement(nums) << endl;
     }
 };
